Add simple controller test
