+++ api/java/src/main/java/com/brandonlind/vendormanagement/service/VendorRepo.java
@@ -1,8 +1,10 @@
-package com.brandonlind.vendormanagement.data;
+package com.brandonlind.vendormanagement.service;
 
 import java.util.UUID;
 
 import org.springframework.data.repository.CrudRepository;
+
+import com.brandonlind.vendormanagement.model.Vendor;
 
 public interface VendorRepo extends CrudRepository<Vendor, UUID> {
   boolean existsByIdAndDeletedAtIsNull(UUID id);
