--- api/java/src/test/java/com/brandonlind/vendormanagement/VendorManagementApplicationTests.java
@@ -1,16 +1,0 @@
-package com.brandonlind.vendormanagement;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class VendorManagementApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
